+++ src/AppIcon/index.ts
@@ -1,0 +1,2 @@
+export { default, icons } from './AppIcon'
+export type { AppIconName, AppIconProps } from './AppIcon'