+++ src/ImageLoader/index.ts
@@ -1,0 +1,2 @@
+export { default } from './ImageLoader'
+export type { ImageLoaderProps } from './ImageLoader'