+++ src/Tabs/index.ts
@@ -1,0 +1,2 @@
+export { default, useTabs } from './Tabs'
+export type { TabItem, TabsProps } from './Tabs'