+++ src/VersionUpdateNotice/index.ts
@@ -1,0 +1,2 @@
+export { DEFAULT_CHECK_INTERVAL_MS, default } from './VersionUpdateNotice'
+export type { VersionUpdateNoticeProps } from './VersionUpdateNotice'